yoav: cleaned smart joystick
