--- src/main/java/frc/utils/joysticks/SmartJoystickConstants.java
@@ -1,15 +1,0 @@
-package frc.utils.joysticks;
-
-class SmartJoystickConstants {
-
-    protected static final double DEADZONE = 0.07;
-
-    protected static final double DEFAULT_THRESHOLD_FOR_AXIS_BUTTON = 0.1;
-
-    protected static final double JOYSTICK_AXIS_TO_SQUARE_FACTOR = Math.sqrt(2);
-
-    protected static final double SENSITIVE_AXIS_VALUE_POWER = 2;
-
-    protected static final double SENSITIVE_SQUARED_AXIS_VALUE_POWER = 2.5;
-
-}
